Change Fibonacci value to be stored in a BigInteger to allow calculations of higher order

+++ app/src/main/java/com/flourenco/fibonacci/core/storage/database/entity/CalculatedFibonacciEntity.kt
@@ -2,10 +2,11 @@ package com.flourenco.fibonacci.core.storage.database.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import java.math.BigInteger
 import org.threeten.bp.LocalDateTime
 
 @Entity
 data class CalculatedFibonacciEntity(
     @PrimaryKey val calculatedNumber: Int,
-    val fibonacciValue: Long
+    val fibonacciValue: BigInteger
 )
