+++ app/src/main/java/com/flourenco/fibonacci/core/storage/database/entity/FibonacciEntryEntity.kt
@@ -2,11 +2,12 @@ package com.flourenco.fibonacci.core.storage.database.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import java.math.BigInteger
 import org.threeten.bp.LocalDateTime
 
 @Entity
 data class FibonacciEntryEntity(
     @PrimaryKey val number: Int,
-    val fibonacciValue: Long,
+    val fibonacciValue: BigInteger,
     val entryTime: LocalDateTime
 )
