+++ app/src/main/java/com/flourenco/fibonacci/model/FibonacciEntry.kt
@@ -1,9 +1,10 @@
 package com.flourenco.fibonacci.model
 
+import java.math.BigInteger
 import org.threeten.bp.LocalDateTime
 
 data class FibonacciEntry(
     val orderNumber: Int,
-    val fibonacciValue: Long,
+    val fibonacciValue: BigInteger,
     val time: LocalDateTime
 )
